% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lateformeeting}
\alias{lateformeeting}
\title{a "dirty" data set to demonstrate rrefine features}
\format{
A data frame with 63 rows and 4 variables
\itemize{
  \item {theDate} {date of observation in varying formats}
  \item {what.day.whas.it} {day of the week in varying formats}
  \item {sleephours} {number of hours slept}
  \item {was.i.on.time.for.work} {indicator of on-time arrival to work}
}
}
\usage{
lateformeeting
}
\description{
This data is a simulated collection of dates, days of the week, numbers of hours slept and indicators of whether or not the subject was on time for work. All observations appearing in this data set are fictitious, and any resemblance to actual arrival times for work is purely coincidental.
}
\examples{
head(lateformeeting)
}
\keyword{datasets}
