% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_check}
\alias{refine_check}
\title{helper function to check if rrefine can connect to OpenRefine}
\usage{
refine_check()
}
\value{
error message if rrefine is unable to connect to OpenRefine, otherwise is invisible
}
\description{
helper function to check if rrefine can connect to OpenRefine
}
