% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_delete.R
\name{refine_delete}
\alias{refine_delete}
\title{delete project from open refine}
\usage{
refine_delete(project.id)
}
\arguments{
\item{project.id}{unique identifier for open refine project to be deleted}
}
\description{
delete project from open refine
}
\examples{
\dontrun{
refine_delete(project.id = 1901017388690)
}

}

