% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_export.R
\name{refine_export}
\alias{refine_export}
\title{export data from open refine}
\usage{
refine_export(project.id, format = "csv", col.names = TRUE,
  encoding = "UTF-8")
}
\arguments{
\item{project.id}{unique identifier for open refine project to be exported}

\item{format}{file format of open refine project to be exported, default is 'csv'}

\item{col.names}{boolean indicator for whether column names should be included.}

\item{encoding}{character encoding for exported data, default is 'UTF-8'}
}
\description{
export data from open refine
}
\examples{
\dontrun{
refine_export(project.id = 1901017388690, format = "csv")
}

}

