% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_operations.R
\name{refine_operations}
\alias{refine_operations}
\title{Title}
\usage{
refine_operations(project.name = NULL, project.id = NULL, operations, ...)
}
\arguments{
\item{project.name}{Name of project to be exported}

\item{project.id}{Unique identifier for project to be exported}

\item{operations}{List}

\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify `host` and `port` arguments if the OpenRefine instance is running at a location other than `http://127.0.0.1:3333`}
}
\value{

}
\description{
Title
}
