% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_operations.R
\name{refine_rename}
\alias{refine_rename}
\title{Rename a column in OpenRefine project}
\usage{
refine_rename(
  original_name,
  new_name,
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{original_name}{Original name for the column}

\item{new_name}{New name for the column}

\item{project.name}{Name of project}

\item{project.id}{Unique identifier for project}

\item{verbose}{Logical specifying whether or not query result should be printed; default is \code{FALSE}}

\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify \code{host} and \code{port} arguments if the OpenRefine instance is running at a location other than \verb{http://127.0.0.1:3333}}
}
\value{
Operates as a side-effect passing operations to the OpenRefine instance. However, if \code{verbose=TRUE} then the function will return an object of the class "response".
}
\description{
This function allows users to rename an existing column in an OpenRefine project via an API query to \verb{/command/core/apply-operations} and the \code{core/column-rename} operation.
}
\examples{

#' \dontrun{
fp <- system.file("extdata", "lateformeeting.csv", package = "rrefine")
refine_upload(fp, project.name = "lfm")
refine_rename("what day whas it", "what_day_was_it", project.name = "lfm")
}


}
