% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_token}
\alias{refine_token}
\title{Helper function to retrieve CSFR token}
\usage{
refine_token()
}
\value{
Character vector with OpenRefine CSFR token
}
\description{
Helper function to retrieve CSFR token
}
