% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_operations.R
\name{transform}
\alias{transform}
\alias{refine_transform}
\alias{refine_to_lower}
\alias{refine_to_upper}
\alias{refine_to_title}
\alias{refine_to_null}
\alias{refine_to_empty}
\alias{refine_to_text}
\alias{refine_to_number}
\alias{refine_to_date}
\title{Text transformation for OpenRefine project}
\usage{
refine_transform(
  column_name,
  expression,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_lower(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_upper(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_title(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_null(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_empty(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_text(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_number(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)

refine_to_date(
  column_name,
  mode = "row-based",
  on_error = "set-to-blank",
  description = "",
  project.name = NULL,
  project.id = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{column_name}{Name of the column on which text transformation should be performed}

\item{expression}{Expression defining the text transformation to be performed}

\item{mode}{Mode of operation; must be one of \code{"row-based"} or \code{"record-based"}; default is \verb{"row-based}}

\item{on_error}{Behavior if there is an error on new column creation; must be one of \code{"set-to-blank"}, \code{"keep-original"}, or \code{"store-error"}; default is \code{"set-to-blank"}}

\item{description}{Text describing the operation performed}

\item{project.name}{Name of project}

\item{project.id}{Unique identifier for project}

\item{verbose}{Logical specifying whether or not query result should be printed; default is \code{FALSE}}

\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify \code{host} and \code{port} arguments if the OpenRefine instance is running at a location other than \verb{http://127.0.0.1:3333}}
}
\value{
Operates as a side-effect passing operations to the OpenRefine instance. However, if \code{verbose=TRUE} then the function will return an object of the class "response".
}
\description{
Text transformation for OpenRefine project
}
