% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_operations.R
\name{refine_add_column}
\alias{refine_add_column}
\title{Title}
\usage{
refine_add_column(
  new_column,
  new_column_index,
  base_column,
  value,
  on_error = "set-to-blank",
  project.name = NULL,
  project.id = NULL,
  ...
)
}
\arguments{
\item{project.name}{Name of project to be exported}

\item{project.id}{Unique identifier for project to be exported}

\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify `host` and `port` arguments if the OpenRefine instance is running at a location other than `http://127.0.0.1:3333`}

\item{column}{Name of the column to be removed}
}
\value{

}
\description{
Title
}
