% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_check}
\alias{refine_check}
\title{Helper function to check if \code{rrefine} can connect to OpenRefine}
\usage{
refine_check()
}
\value{
Error message if \code{rrefine} is unable to connect to OpenRefine, otherwise is invisible
}
\description{
This function will check that \code{rrefine} is able to access the running OpenRefine instance. Used internally prior to upload, delete, and export operations.
}
