% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_check}
\alias{refine_check}
\title{Helper function to check if \code{rrefine} can connect to OpenRefine}
\usage{
refine_check()
}
\value{
Error message if \code{rrefine} is unable to connect to OpenRefine, otherwise is invisible
}
\description{
Helper function to check if \code{rrefine} can connect to OpenRefine
}
