% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_delete.R
\name{refine_delete}
\alias{refine_delete}
\title{delete project from OpenRefine}
\usage{
refine_delete(project.name = NULL, project.id = NULL)
}
\arguments{
\item{project.name}{name of project to be deleted}

\item{project.id}{unique identifier for open refine project to be deleted}
}
\description{
delete project from OpenRefine
}
\examples{
\dontrun{
refine_delete(project.name = "foo")
refine_delete(project.id = 1901017388690)
refine_delete(project.name = "Untitled", project.id = 1901018888332)
}

}

