% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_id}
\alias{refine_id}
\title{Helper function to get OpenRefine project.id by project.name}
\usage{
refine_id(project.name, project.id)
}
\arguments{
\item{project.name}{Name of project}

\item{project.id}{Unique identifier for project}
}
\value{
Unique id of project
}
\description{
Helper function to get OpenRefine project.id by project.name
}
