% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_id}
\alias{refine_id}
\title{helper function to get OpenRefine project.id by project.name}
\usage{
refine_id(project.name, project.id)
}
\arguments{
\item{project.name}{name of project to be exported or deleted}

\item{project.id}{unique identifier for project to be exported or deleted}
}
\value{
unique id of project to be exported or deleted
}
\description{
helper function to get OpenRefine project.id by project.name
}
