% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_metadata}
\alias{refine_metadata}
\title{Get all project metadata from OpenRefine}
\usage{
refine_metadata(...)
}
\arguments{
\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify \code{host} and \code{port} arguments if the OpenRefine instance is running at a location other than \verb{http://127.0.0.1:3333}}
}
\value{
Parsed \code{list} object with all project metadata including identifiers, names, dates of creation and modification, tags and more.
}
\description{
This function is included internally to help retrieve metadata from the running OpenRefine instance. The query uses the OpenRefine API \verb{/command/core/get-all-project-metadata} endpoint.
}
\examples{
\dontrun{
refine_metadata()
}

}
\references{
\url{https://github.com/OpenRefine/OpenRefine/wiki/OpenRefine-API#get-all-projects-metadata}
}
