% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_path}
\alias{refine_path}
\title{Helper function to configure and call path to OpenRefine}
\usage{
refine_path(host = "127.0.0.1", port = "3333")
}
\arguments{
\item{host}{Host for running OpenRefine instance; default is `http://127.0.0.1`}

\item{port}{Port number for running OpenRefine instance; default is `3333`}
}
\value{
Path to running
}
\description{
Helper function to configure and call path to OpenRefine
}
