% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_path}
\alias{refine_path}
\title{helper function to configure and call path to open refine}
\usage{
refine_path(host = "127.0.0.1", port = "3333")
}
\arguments{
\item{host}{host path for your open refine application}

\item{port}{port number for your open refine application}
}
\value{
path to be executed
}
\description{
helper function to configure and call path to open refine
}

