% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_path}
\alias{refine_path}
\title{Helper function to configure and call path to OpenRefine}
\usage{
refine_path(host = "http://127.0.0.1", port = "3333")
}
\arguments{
\item{host}{Host for running OpenRefine instance; default is \code{http://127.0.0.1}}

\item{port}{Port number for running OpenRefine instance; default is \code{3333}}
}
\value{
Character vector with path to running OpenRefine instance
}
\description{
This function is a helper that is used throughout \code{rrefine} to construct the path to the OpenRefine instance. By default this points to the localhost (\code{http://127.0.0.1:3333}).
}
