% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_query}
\alias{refine_query}
\title{Helper function to build OpenRefine API query}
\usage{
refine_query(query, use_token = TRUE)
}
\arguments{
\item{query}{Character vector specifying the API endpoint to query}

\item{use_token}{Boolean indicating whether or not the query string should include a CSRF Token (see \code{\link{refine_token}}; default is \code{TRUE}}
}
\value{
Character vector with query based on parameter entered
}
\description{
Helper function to build OpenRefine API query
}
