% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_token}
\alias{refine_token}
\title{Helper function to retrieve CSFR token}
\usage{
refine_token(...)
}
\arguments{
\item{...}{Additional parameters to be inherited by \code{\link{refine_path}}; allows users to specify \code{host} and \code{port} arguments if the OpenRefine instance is running at a location other than \code{http://127.0.0.1:3333}}
}
\value{
Character vector with OpenRefine CSFR token
}
\description{
Helper function to retrieve CSFR token
}
