% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_upload.R
\name{refine_upload}
\alias{refine_upload}
\title{helper function to configure and call path to open refine}
\usage{
refine_upload(file, project_name, open.browser = FALSE)
}
\arguments{
\item{file}{file name to be uploaded}

\item{project_name}{name of the project to be created upon upload}

\item{open.browser}{boolean for whether or not you want to open browser}
}
\description{
helper function to configure and call path to open refine
}
\examples{
\dontrun{
write.csv(x = mtcars, file = "mtcars.csv")
refine_upload(file = "mtcars.csv", project_name = "mtcars clean-up")
}

}

