% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_upload.R
\name{refine_upload}
\alias{refine_upload}
\title{upload a csv file to OpenRefine}
\usage{
refine_upload(file, project.name = NULL, open.browser = FALSE)
}
\arguments{
\item{file}{Name of csv file to be uploaded}

\item{project.name}{optional parameter to specify name of the project to be created upon upload, default is NULL and project will be named 'Untitled' in OpenRefine}

\item{open.browser}{logical for whether or not the browser should open on successful upload}
}
\description{
upload a csv file to OpenRefine
}
\examples{
\dontrun{
write.csv(x = mtcars, file = "mtcars.csv")
refine_upload(file = "mtcars.csv", project.name = "mtcars_clean_up")
}

}
